+++ PlanetDotnet/Properties/serviceDependencies.json
